backend/requirements.txt
fastapi[standard]==0.115.6
websockets==14.1
uvicorn==0.34.0
pymongo==4.10.1
redis==5.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.4
python-multipart==0.0.20